Mohamed branch

+++ sql_scripts/create_course_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE course (
     description VARCHAR(255) NOT NULL,
     category ENUM('Beginner', 'Intermediate', 'Professional') NOT NULL,
     instructor_name VARCHAR(255) NOT NULL,
-    rating INT NOT NULL,
+    rating DOUBLE NOT NULL,
     duration FLOAT NOT NULL,
     content VARCHAR(255) NOT NULL
 );
